Refactor to use consistent route handling pattern
